+++ src/deluge/model/clip/instrument_clip.h
@@ -76,6 +76,8 @@ public:
 	void seeWhatNotesWithinOctaveArePresent(bool[], int32_t, Song* song, bool deleteEmptyNoteRows = true);
 	void transpose(int32_t, ModelStackWithTimelineCounter* modelStack);
 	void nudgeNotesVertically(int32_t, ModelStackWithTimelineCounter* modelStack);
+	int32_t semitonesNudged;
+	void displayChangeInSemitones();
 	void expectNoFurtherTicks(Song* song, bool actuallySoundChange = true);
 	int32_t clone(ModelStackWithTimelineCounter* modelStack, bool shouldFlattenReversing = false);
 	NoteRow* createNewNoteRowForYVisual(int32_t, Song* song);